Made the full working hologram UI
Hologram UI now has a 'scanner' that displays the info with a small blueish box that displays speed, status, and a bar for growth / hunger that moves as animals hunger is satiated or plants grow more. The hologram has a soft glow on the edges and uses a particle system to look like its kind of flashing.